Add new video channel and resources for 'ジョブレイバー'
Updated the createContentsStrings.sh script to include processing for the new 'ジョブレイバー' channel. Added new video content and images related to 'ジョブレイバー' in the project. Introduced a custom font extension for better font management and added necessary entitlements for iOS. Updated Info.plist and project settings to accommodate new resources and functionalities.